requirements_simple.txt
streamlit==1.31.0
pypdf==4.0.1
tiktoken==0.6.0
anthropic==0.40.0
voyageai==0.2.1
python-dotenv==1.0.0
numpy==1.26.4
